requirements.txt
Flask==3.1.1
gunicorn
cloudinary
python-dotenv